+++ empty-sidebars.php
@@ -5,7 +5,7 @@ Template Name: Empty Sidebars
 ?>
 <?php get_header(); ?>
 
-<div class="container">
+<div class="container" role="content" aria_label="Main Content">
   <div class="row">
 
     <?php $has_left = FALSE; ?>